Move gmsh finalize.
